% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyzeEnrichment.R
\name{analyzeEnrichment}
\alias{analyzeEnrichment}
\title{Title}
\usage{
analyzeEnrichment(regionPath, regionHeader = c("chr", "start", "end"),
  SNPsnapPath, numberOfRandomSNPsets, LSProxyPathPrefix, BGProxyPathPrefix,
  traitShort, genomicRegionsName, cores, resDir, traitsLong = NULL)
}
\arguments{
\item{regionPath}{}

\item{regionHeader}{}

\item{SNPsnapPath}{}

\item{numberOfRandomSNPsets}{}

\item{traitShort}{}

\item{genomicRegionsName}{}

\item{resDir}{}

\item{traitsLong}{}

\item{proxyPathPrefix}{}
}
\description{
Title
}
