% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpSNPs.R
\name{clumpSNPs}
\alias{clumpSNPs}
\title{Title}
\usage{
clumpSNPs(plinkPathPrefix, SNPsPath, outputPath, corrCoeff, distance)
}
\arguments{
\item{plinkPathPrefix}{}

\item{SNPsPath}{}

\item{outputPath}{}

\item{corrCoeff}{}

\item{distance}{}
}
\description{
Title
}
