% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichTest.R
\name{enrichTest}
\alias{enrichTest}
\title{Test for enrichment of SNPs within a given feature list.}
\usage{
enrichTest(matchedSNPfile, proxiesFile, featureObj)
}
\arguments{
\item{matchedSNPfile}{matched_snps.txt file from SNPsnap server}

\item{proxiesFile}{File containing proxies}

\item{featureObj}{Feature object for which enrichment has to be calculated}
}
\value{
Returns p.value
}
\description{
Test for enrichment of SNPs within a given feature list.
}
\examples{
enrichTest()
}
