% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitSNPsnap.R
\name{submitSNPsnap}
\alias{submitSNPsnap}
\title{Submit list of SNPs to SNPsnap server}
\usage{
submitSNPsnap(snplist, super_population = c("EUR", "EAS", "WAFR"),
  distance_type = c("ld", "kb"), distance, max_freq_deviation = seq(1, 50,
  1), max_genes_count_deviation, max_distance_deviation,
  max_ld_buddy_count_deviation, ld_buddy_cutoff = seq(0.1, 0.9, 0.1),
  N_sample_sets, annotate_matched = FALSE, annotate_input = FALSE,
  clump_input = TRUE, clump_r2 = seq(0.1, 0.9, 0.1), clump_kb = seq(100,
  1000, 100), exclude_input_SNPs = TRUE, exclude_HLA_SNPs = TRUE,
  email_address, job_name)
}
\arguments{
\item{snplist}{List of SNPs}

\item{super_population}{Super population to use; defaults to 'EUR'}

\item{distance_type}{Distance type; default is 'ld'}

\item{distance}{Distance to use}

\item{max_freq_deviation}{Percentage deviation from query SNP minor allele frequency}

\item{max_genes_count_deviation}{Gene density deviation from query SNP}

\item{max_distance_deviation}{Gene distance deviation from query SNP}

\item{max_ld_buddy_count_deviation}{Deviation of number of LD buddies near query SNP}

\item{ld_buddy_cutoff}{Cutoff of R2 while calculation LD buddies}

\item{N_sample_sets}{Number of SNP sets requested}

\item{annotate_matched}{a logical value to annotate matched SNPs}

\item{annotate_input}{a logical value to annotate input SNPs}

\item{clump_input}{a logical value to report clumping of input SNPs}

\item{clump_r2}{R2 cutoff for clumping}

\item{clump_kb}{Distance cutoff for clumping}

\item{exclude_input_SNPs}{a logical value to exclude input SNPs in matched SNPs}

\item{exclude_HLA_SNPs}{a logical value to exclude HLA SNPs}

\item{email_address}{Your email address}

\item{job_name}{Job name}
}
\value{
Returns a URL from which results can be downloaded
}
\description{
SubmitSNPsnap submits a query to SNPsnap server for creating random sets of SNPs matching input SNPs and 
the criteria given by the user.
}
\examples{
submitSNPsnap(snplist=snps,email_address='kartiek.kanduri@aalto.fi')
}
\author{
Kartiek Kanduri, Kari Nousiainen
}
