% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpSNPs.R
\name{clumpSNPs}
\alias{clumpSNPs}
\title{Decipher a set of non-clumped SNPs from a list of SNPs}
\usage{
clumpSNPs(plinkPathPrefix, snplist, outputPath, clump_r2, clump_kb)
}
\arguments{
\item{plinkPathPrefix}{Path to reference directory, must contain the prefix of the plink reference files bed, bim and fam.}

\item{snplist}{Character vector of snps.}

\item{outputPath}{A character string is the Path to output file.}

\item{clump_r2}{R2 cutoff for clumping.}

\item{clump_kb}{Distance cutoff for clumping.}
}
\value{
Returns a character vector of decorrelated SNPs.
}
\description{
ClumpSNPs is a wrapper to plink 1.9. The function takes a character vector of SNPs and 
computes the clumped SNP in the vector. The output contains all SNPs without clumping and one 
representative for every clump.
}
\examples{
snpList <- clumpSNPs(plinkPathPrefix,snplist,outputPath,clump_r2,clump_kb)
}
