% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpSNPs.R
\name{clumpSNPs}
\alias{clumpSNPs}
\title{Decipher a set of non-clumped SNPs from a list of SNPs}
\usage{
clumpSNPs(plinkPathPrefix, snplist, outputPath, clump_r2, clump_kb)
}
\arguments{
\item{plinkPathPrefix}{Path to reference directory, must contain the prefix of the plink reference files bed, bim and fam. For preprocessed files, see library(snpEnrichR)}

\item{snplist}{Character vector of snps.}

\item{outputPath}{A character string is the Path to output file.}

\item{clump_r2}{R2 cutoff for clumping.}

\item{clump_kb}{Distance cutoff for clumping.}
}
\value{
Returns a character vector of decorrelated SNPs.
}
\description{
ClumpSNPs is a wrapper to plink 1.9. The function takes a character vector of SNPs and 
computes the clumped SNP in the vector. The output contains all SNPs without clumping and one 
representative for every clump.
}
\examples{
snpList <- clumpSNPs(plinkPathPrefix,snplist,outputPath,clump_r2,clump_kb)
}
\references{
Chang, C. C., Chow, C. C., Tellier, L. C., Vattikuti, S., Purcell, S. M., Lee, J. J. (2015). Second-generation PLINK: rising to the challenge of larger and richer datasets. \emph{Gigascience}, 4(1), 7.

Purcell S.M., Chang C.C. \emph{PLINK 1.9} \url{ www.cog-genomics.org/plink/1.9/}.
}
