% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clumpSNPs.R
\name{clumpSNPs}
\alias{clumpSNPs}
\title{Decipher a set of non-clumped SNPs from a list of SNPs}
\usage{
clumpSNPs(plinkPathPrefix, snplist, outputPath, clump_r2, clump_kb)
}
\arguments{
\item{plinkPathPrefix}{Path to reference directory, must contain the prefix of the plink reference files bed, bim and fam.}

\item{snplist}{Character vector of snps.}

\item{outputPath}{Path to output file.}

\item{clump_r2}{R2 cutoff for clumping.}

\item{clump_kb}{Distance cutoff for clumping.}
}
\value{
Returns SNP list with clumped SNP buddies
}
\description{
Creates a list of non-clumped snps with one representative of each clump in the input snp list and writes it to a file.
}
\examples{
snpList <- clumpSNPs(plinkPathPrefix,snplist,outputPath,clump_r2,clump_kb)
}
