% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProxies.R
\name{findProxies}
\alias{findProxies}
\title{Find proxies to SNPs using plink}
\usage{
findProxies(path2PlinkPrefix, snplist, proxyWindow = 1000, proxyCorr = 0.8,
  path2Proxies)
}
\arguments{
\item{path2PlinkPrefix}{Path to reference directory, must contain the prefix of the plink reference files bed, bim and fam}

\item{proxyWindow}{--ld-window-kb;  default is 1000}

\item{proxyCorr}{plink parameter --ld-window-r2 ; default is 0.8}

\item{path2Proxies}{path to directory where the SNPs are}

\item{path2leadSNPList}{Full path to the list of lead snps}
}
\description{
Computes proxies for given SNPs using plink
}
\examples{
findProxies(path2PlinkPrefix, snplist, path2Proxies)
}
\author{
Kari Nousiainen
}
