% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProxies.R
\name{findProxies}
\alias{findProxies}
\title{Submit list of SNPs to SNPsnap server}
\usage{
findProxies(path2RefDir, path2leadSNPList, proxyWindow = 1000,
  proxyCorr = 0.8, path2Proxies, useMaf = TRUE, maf = 0.01,
  useHWE = TRUE, hwe = 1e-05, useGENO = TRUE, geno = 0.1)
}
\arguments{
\item{path2RefDir}{Path to reference directory, must contain the prefix of the plink reference files bed, bim and fam}

\item{path2leadSNPList}{Full path to the list of lead snps}

\item{proxyWindow}{--ld-window-kb;  default is 1000}

\item{proxyCorr}{plink parameter --ld-window-r2 ; default is 0.8}

\item{path2Proxies}{patht to directory where the SNPs are}

\item{useMaf, }{a logical value to decide whether maf pameter is used;  default is TRUE}

\item{maf}{plink parameter maf;  default is 0.01}

\item{useHWE, }{a logical value to decide whether HWE pameter is used;  default is TRUE}

\item{hwe}{plink parameter hwe;  default is 10e-6}

\item{useGENO, }{a logical value to decide whether geno pameter is used;  default is TRUE}

\item{geno}{plink parameter 0.1; default is  0.1}

\item{referenceDir}{}
}
\description{
Submit list of SNPs to SNPsnap server
}
\examples{
findProxies()
}
