% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findProxies.R
\name{findProxies}
\alias{findProxies}
\title{Find proxies to SNPs using plink}
\usage{
findProxies(path2PlinkPrefix, snplist, ld_window_kb = 1000,
  ld_window_r2 = 0.8, path2Proxies)
}
\arguments{
\item{path2PlinkPrefix}{Plink bfile parameter. It is the path to reference directory including prefix of the plink reference files bed, bim and fam}

\item{ld_window_kb}{Plink parameter ld_window_kb denotes the maximum distance between LD buddies (default is 1000).}

\item{ld_window_r2}{Plink parameter ld_window_r2 denoted the minimum correlation  between LD buddies (default is 0.8).}

\item{path2Proxies}{Path to directory where the result will be written.}

\item{path2leadSNPList}{Full path to the SNP list}
}
\description{
ClumpSNPs is a wrapper to plink 1.9. The function computes proxy SNPs for the input snps 
and writes them to a file.
}
\examples{
findProxies(path2PlinkPrefix, snplist, path2Proxies)
}
\author{
Kari Nousiainen
}
