% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPs.R
\name{getSNPs}
\alias{getSNPs}
\title{Get list of SNPs from EBI's GWAS server}
\usage{
getSNPs(queryTerm)
}
\arguments{
\item{queryTerm}{The query term for EBI GWAS. See https://www.ebi.ac.uk/gwas/ for details.}
}
\value{
Returns a data frame of SNPs from EBI's server
}
\description{
GetSNPs takes a character string and sends it to NHGRI-EBI GWAS Catalog as the query term 
and returns a data table containing all data matching to the query term.
}
\examples{
getSNPs('breast cancer')
}
