% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSNPs.R
\name{getSNPs}
\alias{getSNPs}
\title{Get list of SNPs from EBI's GWAS server}
\usage{
getSNPs(queryTerm)
}
\arguments{
\item{queryTerm}{Query term for EBI GWAS. See https://www.ebi.ac.uk/gwas/ for details.}
}
\value{
Returns a data fram of SNPs from EBI's server
}
\description{
Get list of SNPs from EBI's GWAS server
}
\examples{
getSNPs('breast cancer')
}
